crud leave done

+++ app/src/main/java/com/example/rapiertech/adapter/AdapterDataDepartment.java
@@ -12,7 +12,6 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AlertDialog;
-import androidx.core.content.res.ResourcesCompat;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.rapiertech.R;
@@ -21,6 +20,7 @@ import com.example.rapiertech.api.ApiInterface;
 import com.example.rapiertech.model.department.Department;
 import com.example.rapiertech.model.department.DepartmentData;
 import com.example.rapiertech.ui.admin.DepartmentFragment;
+import com.example.rapiertech.widget.Widget;
 import com.google.android.material.dialog.MaterialAlertDialogBuilder;
 
 import org.jetbrains.annotations.NotNull;
@@ -31,13 +31,13 @@ import dev.shreyaspatil.MaterialDialog.BottomSheetMaterialDialog;
 import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
-import www.sanju.motiontoast.MotionToast;
 
 public class AdapterDataDepartment extends RecyclerView.Adapter<AdapterDataDepartment.HolderData>{
     private final Context context;
     private final List<DepartmentData> listDept;
     private String name;
     private final DepartmentFragment fragment;
+    private Widget widget;
 
     public AdapterDataDepartment(Context context, List<DepartmentData> listDept, DepartmentFragment fragment) {
         this.context = context;
@@ -56,56 +56,52 @@ public class AdapterDataDepartment extends RecyclerView.Adapter<AdapterDataDepar
     @Override
     public void onBindViewHolder(@NonNull @NotNull HolderData holder, int position) {
         DepartmentData dd = listDept.get(position);
-
+        widget = new Widget();
         holder.tvId.setText(String.valueOf(dd.getId()));
         holder.tvName.setText(dd.getName());
         holder.totalDept.setText(dd.getTotal() + " Employee");
         if (dd.getTotal() == 0){
-            holder.totalDept.setBackgroundTintList(context.getColorStateList(R.color.danger));
+            holder.totalDept.setBackgroundTintList(context.getColorStateList(R.color.light_red));
         } else {
-            holder.totalDept.setBackgroundTintList(context.getColorStateList(R.color.success));
+            holder.totalDept.setBackgroundTintList(context.getColorStateList(R.color.light_green));
         }
         holder.menuPopup.setOnClickListener(v -> {
             PopupMenu popupMenu = new PopupMenu(context, v);
             popupMenu.getMenuInflater().inflate(R.menu.home, popupMenu.getMenu());
             popupMenu.show();
             popupMenu.setOnMenuItemClickListener(item -> {
-                switch (item.getItemId())
-                {
-                    case R.id.action_edit:
-                        MaterialAlertDialogBuilder editDialog = new MaterialAlertDialogBuilder(context);
-                        LayoutInflater inflater = LayoutInflater.from(context);
-                        View view = inflater.inflate(R.layout.add_dialog_deptjob, null);
+                int itemId = item.getItemId();
+                if (itemId == R.id.action_edit) {
+                    MaterialAlertDialogBuilder editDialog = new MaterialAlertDialogBuilder(context);
+                    LayoutInflater inflater = LayoutInflater.from(context);
+                    View view = inflater.inflate(R.layout.add_dialog_deptjob, null);
 
-                        EditText etName = view.findViewById(R.id.add_deptJobName);
-                        etName.setText(dd.getName());
+                    EditText etName = view.findViewById(R.id.add_deptJobName);
+                    etName.setText(dd.getName());
 
-                        editDialog.setView(view)
-                                .setTitle(R.string.add_title_dialog)
-                                .setPositiveButton(R.string.save, (dialog, which) -> {
-                                    name = etName.getText().toString();
-                                    updateData(dd.getId());
-                                    dialog.dismiss();
-                                })
-                                .setNegativeButton(R.string.cancel, (dialog, which) -> dialog.dismiss());
+                    editDialog.setView(view)
+                            .setTitle(R.string.add_title_dialog)
+                            .setPositiveButton(R.string.save, (dialog, which) -> {
+                                name = etName.getText().toString();
+                                updateData(dd.getId());
+                                dialog.dismiss();
+                            })
+                            .setNegativeButton(R.string.cancel, (dialog, which) -> dialog.dismiss());
 
-                        AlertDialog alertDialog = editDialog.create();
-                        alertDialog.show();
-                        break;
-
-                    case R.id.action_delete:
-                        BottomSheetMaterialDialog mDialog = new BottomSheetMaterialDialog.Builder((Activity) context)
-                                .setTitle("Delete?")
-                                .setMessage("Are you sure want to delete this data?")
-                                .setCancelable(false)
-                                .setPositiveButton("Delete", R.drawable.ic_delete_, (dialogInterface, which) -> {
-                                    deleteData(dd.getId());
-                                    dialogInterface.dismiss();
-                                })
-                                .setNegativeButton("Cancel", R.drawable.ic_close, (dialogInterface, which) -> dialogInterface.dismiss())
-                                .build();
-                        mDialog.show();
-                        break;
+                    AlertDialog alertDialog = editDialog.create();
+                    alertDialog.show();
+                } else if (itemId == R.id.action_delete) {
+                    BottomSheetMaterialDialog mDialog = new BottomSheetMaterialDialog.Builder((Activity) context)
+                            .setTitle("Delete?")
+                            .setMessage("Are you sure want to delete this data?")
+                            .setCancelable(false)
+                            .setPositiveButton("Delete", R.drawable.ic_delete_, (dialogInterface, which) -> {
+                                deleteData(dd.getId());
+                                dialogInterface.dismiss();
+                            })
+                            .setNegativeButton("Cancel", R.drawable.ic_close, (dialogInterface, which) -> dialogInterface.dismiss())
+                            .build();
+                    mDialog.show();
                 }
                 return true;
             });
@@ -118,17 +114,19 @@ public class AdapterDataDepartment extends RecyclerView.Adapter<AdapterDataDepar
         updatedata.enqueue(new Callback<Department>() {
             @Override
             public void onResponse(Call<Department> call, Response<Department> response) {
-                if (response.body() != null && response.isSuccessful() && response.body().isStatus()){
-                    successToast(response.body().getMessage());
-                    fragment.retrieveData();
-                }else{
-                    errorToast(response.body().getMessage());
+                if (response.body() != null){
+                    if (response.isSuccessful() && response.body().isStatus()){
+                        widget.successToast(response.body().getMessage(), fragment.requireActivity());
+                        fragment.retrieveData();
+                    }else{
+                        widget.errorToast(response.body().getMessage(), fragment.requireActivity());
+                    }
                 }
             }
 
             @Override
             public void onFailure(Call<Department> call, Throwable t) {
-                noConnectToast(t.getMessage());
+                widget.noConnectToast(t.getMessage(), fragment.requireActivity());
             }
         });
     }
@@ -139,17 +137,19 @@ public class AdapterDataDepartment extends RecyclerView.Adapter<AdapterDataDepar
         deletedata.enqueue(new Callback<Department>() {
             @Override
             public void onResponse(Call<Department> call, Response<Department> response) {
-                if (response.isSuccessful()){
-                    successToast(response.body().getMessage());
-                    fragment.retrieveData();
-                } else {
-                    errorToast(response.body().getMessage());
+                if (response.body() != null) {
+                    if (response.isSuccessful() && response.body().isStatus()) {
+                        widget.successToast(response.body().getMessage(), fragment.requireActivity());
+                        fragment.retrieveData();
+                    } else {
+                        widget.errorToast(response.body().getMessage(), fragment.requireActivity());
+                    }
                 }
             }
 
             @Override
             public void onFailure(Call<Department> call, Throwable t) {
-                noConnectToast(t.getMessage());
+                widget.noConnectToast(t.getMessage(), fragment.requireActivity());
             }
         });
     }
@@ -171,35 +171,5 @@ public class AdapterDataDepartment extends RecyclerView.Adapter<AdapterDataDepar
             menuPopup = itemView.findViewById(R.id.more_deptJob);
             totalDept = itemView.findViewById(R.id.total_deptJob);
         }
-    }
-
-    private void noConnectToast(String message) {
-        MotionToast.Companion.createColorToast((Activity) context, "Cannot connect server",
-                message,
-                MotionToast.TOAST_ERROR,
-                MotionToast.GRAVITY_BOTTOM,
-                MotionToast.LONG_DURATION,
-                ResourcesCompat.getFont(context,R.font.helvetica_regular)
-        );
-    }
-
-    private void errorToast(String message) {
-        MotionToast.Companion.createColorToast((Activity) context, "Error",
-                message,
-                MotionToast.TOAST_ERROR,
-                MotionToast.GRAVITY_BOTTOM,
-                MotionToast.LONG_DURATION,
-                ResourcesCompat.getFont(context,R.font.helvetica_regular)
-        );
-    }
-
-    private void successToast(String message) {
-        MotionToast.Companion.createColorToast((Activity) context, "Success",
-                message,
-                MotionToast.TOAST_SUCCESS,
-                MotionToast.GRAVITY_BOTTOM,
-                MotionToast.LONG_DURATION,
-                ResourcesCompat.getFont(context,R.font.helvetica_regular)
-        );
     }
 }
