+++ app/src/main/java/com/example/rapiertech/adapter/AdapterDataJob.java
@@ -13,7 +13,6 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AlertDialog;
-import androidx.core.content.res.ResourcesCompat;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.rapiertech.R;
@@ -22,6 +21,7 @@ import com.example.rapiertech.api.ApiInterface;
 import com.example.rapiertech.model.job.Job;
 import com.example.rapiertech.model.job.JobData;
 import com.example.rapiertech.ui.admin.JobFragment;
+import com.example.rapiertech.widget.Widget;
 import com.google.android.material.dialog.MaterialAlertDialogBuilder;
 
 import org.jetbrains.annotations.NotNull;
@@ -32,7 +32,6 @@ import dev.shreyaspatil.MaterialDialog.BottomSheetMaterialDialog;
 import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
-import www.sanju.motiontoast.MotionToast;
 
 public class AdapterDataJob extends RecyclerView.Adapter<AdapterDataJob.HolderData> {
 
@@ -40,6 +39,7 @@ public class AdapterDataJob extends RecyclerView.Adapter<AdapterDataJob.HolderDa
     private final List<JobData> jobList;
     private String name;
     private final JobFragment fragment;
+    private Widget widget;
 
     public AdapterDataJob(Context context, List<JobData> jobList, JobFragment fragment){
         this.context = context;
@@ -58,14 +58,15 @@ public class AdapterDataJob extends RecyclerView.Adapter<AdapterDataJob.HolderDa
     @Override
     public void onBindViewHolder(@NonNull @NotNull HolderData holder, int position) {
         JobData jd = jobList.get(position);
+        widget = new Widget();
 
         holder.tvId.setText(String.valueOf(jd.getId()));
         holder.tvName.setText(jd.getName());
         holder.totalDept.setText(jd.getTotal() + " Employee");
         if (jd.getTotal() == 0){
-            holder.totalDept.setBackgroundTintList(context.getColorStateList(R.color.danger));
+            holder.totalDept.setBackgroundTintList(context.getColorStateList(R.color.light_red));
         } else {
-            holder.totalDept.setBackgroundTintList(context.getColorStateList(R.color.success));
+            holder.totalDept.setBackgroundTintList(context.getColorStateList(R.color.light_green));
         }
         holder.menuPopup.setOnClickListener(v -> {
             PopupMenu popupMenu = new PopupMenu(context, v);
@@ -75,42 +76,38 @@ public class AdapterDataJob extends RecyclerView.Adapter<AdapterDataJob.HolderDa
                 popupMenu.setForceShowIcon(true);
             }
             popupMenu.setOnMenuItemClickListener(item -> {
-                switch (item.getItemId())
-                {
-                    case R.id.action_edit:
-                        MaterialAlertDialogBuilder editDialog = new MaterialAlertDialogBuilder(context);
-                        LayoutInflater inflater = LayoutInflater.from(context);
-                        View view = inflater.inflate(R.layout.add_dialog_deptjob, null);
+                int itemId = item.getItemId();
+                if (itemId == R.id.action_edit) {
+                    MaterialAlertDialogBuilder editDialog = new MaterialAlertDialogBuilder(context);
+                    LayoutInflater inflater = LayoutInflater.from(context);
+                    View view = inflater.inflate(R.layout.add_dialog_deptjob, null);
 
-                        EditText etName = view.findViewById(R.id.add_deptJobName);
-                        etName.setText(jd.getName());
+                    EditText etName = view.findViewById(R.id.add_deptJobName);
+                    etName.setText(jd.getName());
 
-                        editDialog.setView(view)
-                                .setTitle(R.string.add_title_dialog)
-                                .setPositiveButton(R.string.save, (dialog, which) -> {
-                                    name = etName.getText().toString();
-                                    updateData(jd.getId());
-                                    dialog.dismiss();
-                                })
-                                .setNegativeButton(R.string.cancel, (dialog, which) -> dialog.dismiss());
+                    editDialog.setView(view)
+                            .setTitle(R.string.add_title_dialog)
+                            .setPositiveButton(R.string.save, (dialog, which) -> {
+                                name = etName.getText().toString();
+                                updateData(jd.getId());
+                                dialog.dismiss();
+                            })
+                            .setNegativeButton(R.string.cancel, (dialog, which) -> dialog.dismiss());
 
-                        AlertDialog alertDialog = editDialog.create();
-                        alertDialog.show();
-                        break;
-
-                    case R.id.action_delete:
-                        BottomSheetMaterialDialog mDialog = new BottomSheetMaterialDialog.Builder((Activity) context)
-                                .setTitle("Delete?")
-                                .setMessage("Are you sure want to delete this data?")
-                                .setCancelable(false)
-                                .setPositiveButton("Delete", R.drawable.ic_delete_, (dialogInterface, which) -> {
-                                    deleteData(jd.getId());
-                                    dialogInterface.dismiss();
-                                })
-                                .setNegativeButton("Cancel", R.drawable.ic_close, (dialogInterface, which) -> dialogInterface.dismiss())
-                                .build();
-                        mDialog.show();
-                        break;
+                    AlertDialog alertDialog = editDialog.create();
+                    alertDialog.show();
+                } else if (itemId == R.id.action_delete) {
+                    BottomSheetMaterialDialog mDialog = new BottomSheetMaterialDialog.Builder((Activity) context)
+                            .setTitle("Delete?")
+                            .setMessage("Are you sure want to delete this data?")
+                            .setCancelable(false)
+                            .setPositiveButton("Delete", R.drawable.ic_delete_, (dialogInterface, which) -> {
+                                deleteData(jd.getId());
+                                dialogInterface.dismiss();
+                            })
+                            .setNegativeButton("Cancel", R.drawable.ic_close, (dialogInterface, which) -> dialogInterface.dismiss())
+                            .build();
+                    mDialog.show();
                 }
                 return true;
             });
@@ -123,17 +120,19 @@ public class AdapterDataJob extends RecyclerView.Adapter<AdapterDataJob.HolderDa
         updatedata.enqueue(new Callback<Job>() {
             @Override
             public void onResponse(Call<Job> call, Response<Job> response) {
-                if (response.body() != null && response.isSuccessful() && response.body().isStatus()){
-                    successToast(response.body().getMessage());
-                    fragment.retrieveData();
-                }else{
-                    errorToast(response.body().getMessage());
+                if (response.body() != null){
+                    if (response.isSuccessful() && response.body().isStatus()){
+                        widget.successToast(response.body().getMessage(), fragment.requireActivity());
+                        fragment.retrieveData();
+                    }else{
+                        widget.errorToast(response.body().getMessage(), fragment.requireActivity());
+                    }
                 }
             }
 
             @Override
             public void onFailure(Call<Job> call, Throwable t) {
-                noConnectToast(t.getMessage());
+                widget.noConnectToast(t.getMessage(), fragment.requireActivity());
             }
         });
     }
@@ -144,17 +143,19 @@ public class AdapterDataJob extends RecyclerView.Adapter<AdapterDataJob.HolderDa
         deletedata.enqueue(new Callback<Job>() {
             @Override
             public void onResponse(Call<Job> call, Response<Job> response) {
-                if (response.isSuccessful()){
-                    successToast(response.body().getMessage());
-                    fragment.retrieveData();
-                } else {
-                    errorToast(response.body().getMessage());
+                if (response.body() != null){
+                    if (response.isSuccessful()){
+                        widget.successToast(response.body().getMessage(), fragment.requireActivity());
+                        fragment.retrieveData();
+                    } else {
+                        widget.errorToast(response.body().getMessage(), fragment.requireActivity());
+                    }
                 }
             }
 
             @Override
             public void onFailure(Call<Job> call, Throwable t) {
-                noConnectToast(t.getMessage());
+                widget.noConnectToast(t.getMessage(), fragment.requireActivity());
             }
         });
     }
@@ -176,35 +177,5 @@ public class AdapterDataJob extends RecyclerView.Adapter<AdapterDataJob.HolderDa
             menuPopup = itemView.findViewById(R.id.more_deptJob);
             totalDept = itemView.findViewById(R.id.total_deptJob);
         }
-    }
-
-    private void noConnectToast(String message) {
-        MotionToast.Companion.createColorToast((Activity) context, "Cannot connect server",
-                message,
-                MotionToast.TOAST_ERROR,
-                MotionToast.GRAVITY_BOTTOM,
-                MotionToast.LONG_DURATION,
-                ResourcesCompat.getFont(context,R.font.helvetica_regular)
-        );
-    }
-
-    private void errorToast(String message) {
-        MotionToast.Companion.createColorToast((Activity) context, "Error",
-                message,
-                MotionToast.TOAST_ERROR,
-                MotionToast.GRAVITY_BOTTOM,
-                MotionToast.LONG_DURATION,
-                ResourcesCompat.getFont(context,R.font.helvetica_regular)
-        );
-    }
-
-    private void successToast(String message) {
-        MotionToast.Companion.createColorToast((Activity) context, "Success",
-                message,
-                MotionToast.TOAST_SUCCESS,
-                MotionToast.GRAVITY_BOTTOM,
-                MotionToast.LONG_DURATION,
-                ResourcesCompat.getFont(context,R.font.helvetica_regular)
-        );
     }
 }
